+++ src/main/java/nextstep/courses/domain/cource/Course.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.cource;
 
 import java.time.LocalDateTime;
